Update to long double and add 10 to the X

+++ src/HP11C.h
@@ -9,7 +9,7 @@
 #include "resources.h"
 #include "window.h"
 
-extern double stack[4];
+extern long double stack[4];
 extern std::wstring stackString;
 extern bool inEditMode;
